navigation to another letter
Navigation from one to another, working with positions. Next: showing a modal.
