+++ package.json
@@ -10,8 +10,10 @@
     "@testing-library/react": "^12.1.2",
     "@testing-library/user-event": "^13.5.0",
     "dat.gui": "^0.7.7",
+    "fast-deep-equal": "^3.1.3",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "three": "^0.136.0",
     "three-mesh-bvh": "^0.5.4",
